Struct redis_client::redis::RedisClientAsync

pub struct RedisClientAsync {
    // some fields omitted
}

Methods

impl RedisClientAsync

fn new(host: &'static str, port: &'static str) -> Result<RedisClientAsync, RedisError>

fn exec_redis_command_async<F>(&self, redis_command: &mut RedisCommand, callback: F) -> Result<(), RedisError> where F: Fn(Result<RedisResult, RedisError>), F: Send + 'static

Execute a redis command and call the callback when it is done with the result The return value indicates if the command was successfully launched